feat: add exception handling to newGroup screen

+++ src/storage/group/groupCreate.ts
@@ -2,11 +2,20 @@ import AsyncStorage from '@react-native-async-storage/async-storage';
 
 import { GROUP_COLLECTION } from '@storage/storageConfig';
 
+import { AppError } from '@utils/AppError';
+
 import { groupGetAll } from './groupGetAll';
 
 export async function groupCreate(groupName: string) {
   try {
     const storedGroups = await groupGetAll();
+
+    const groupAlreadyExists = storedGroups.includes(groupName);
+
+    if (groupAlreadyExists) {
+      throw new AppError('Já existe um grupo cadastrado com esse nome.');
+    }
+
     const storage = JSON.stringify([...storedGroups, groupName]);
     await AsyncStorage.setItem(GROUP_COLLECTION, storage);
   } catch (error) {
